+++ CHANGELOG.md
@@ -19,6 +19,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - [#31](https://github.com/oleasteo/koreader-screenlockpin/issues/31)
   Initial bottom config tab change now properly hides prior tab for smaller next
   tab.
+- [#41](https://github.com/oleasteo/koreader-screenlockpin/issues/41)
+  Make sure to show the PIN panel on wakeup in case of lock on wakeup being
+  disabled, but lock screen active elsewise (e.g., dispatcher / lock on boot).
 
 ## [2025.12-1] - 2025-12-22